fix up docker, add doc strings, dev cog. update to dpy 2.0

+++ cogs/management.py
@@ -15,23 +15,33 @@ class ManagementCog(commands.Cog, name="Management"):
 
     @commands.group(name="levels")
     async def levels_group(self, ctx: Context):
+        """
+        The base group for setting up level rewards.
+        """
         if ctx.invoked_subcommand is None:
             await ctx.send_help("levels")
 
     @levels_group.command(name="list")
     async def list_levels(self, ctx: Context):
+        """
+        Lists the currently active roles with their levels.
+        """
         end = ""
 
-        for level, id in self.bot.roles:
+        for level, id in self.bot.roles.items():
             role = ctx.guild.get_role(id)
-            end += f"``**{level}**``: {role.mention} \n"
+            end += f"``{level}``: {role.mention} \n"
 
         await ctx.send(embed=discord.Embed(description=end))
 
     @commands.has_guild_permissions(manage_guild=True)
     @levels_group.command(name="add")
     async def add_role(self, ctx: Context, role: discord.Role, level: int):
-        if self.bot.roles.get(level) == level:
+        """
+        Add a role bound to a level
+        You cannot have multiple roles bound to a single level.
+        """
+        if level in self.bot.roles.keys():
             raise commands.BadArgument("Level cannot already have a role, please use `levels remove <level>` to remove it and try again.")
 
         row = await self.bot.db.fetchone((
@@ -47,11 +57,71 @@ class ManagementCog(commands.Cog, name="Management"):
     @commands.has_guild_permissions(manage_guild=True)
     @levels_group.command(name="remove")
     async def remove_role(self, ctx: Context, level: int):
+        """
+        Removes a role from the role rewards.
+        """
+        if level not in self.bot.roles.keys():
+            raise commands.BadArgument("That level has no role bound to it.")
+
         await self.bot.db.execute(
             "DELETE FROM roles WHERE level = $1", level
         )
 
+        self.bot.roles.pop(level, None)
+
         await ctx.send(embed=discord.Embed(description=f"Successfully removed the role given for level ``**{level}**``"))
+
+    @commands.has_guild_permissions(manage_guild=True)
+    @commands.group(name="blacklist")
+    async def blacklist_group(self, ctx: commands.Context):
+        """
+        Base group for the blacklist command.
+        """
+        if ctx.invoked_subcommand is None:
+            await ctx.send_help("blacklist")
+
+    @commands.has_guild_permissions(manage_guild=True)
+    @blacklist_group.command(name="add")
+    async def add_channel(self, ctx: commands.Context, *, channel: discord.TextChannel):
+        """
+        Add a channel to the blacklist.
+        """
+        if channel.id in self.bot.blacklist:
+            raise commands.BadArgument("This channel is already blacklisted.")
+
+        async with self.bot.db.acquire() as conn:
+            await conn.execute(
+                "INSERT INTO blacklist(id) VALUES($1) RETURNING *", channel.id
+            )
+
+            self.bot.blacklist.append(channel.id)
+
+        await ctx.message.add_reaction("\U00002705")
+
+    @commands.has_guild_permissions(manage_guild=True)
+    @blacklist_group.command(name="remove")
+    async def remove_channel(self, ctx: commands.Context, *, channel: discord.TextChannel):
+        """
+        Remove a channel from the blacklist.
+        """
+        if channel.id not in self.bot.blacklist:
+            raise commands.BadArgument("This channel is not blacklisted.")
+
+        async with self.bot.db.acquire() as conn:
+            await conn.execute(
+                "DELETE FROM blacklist WHERE id = $1", channel.id
+            )
+
+            self.bot.blacklist.remove(channel.id)
+        
+        await ctx.message.add_reaction("\U00002705")
+
+    @blacklist_group.command(name="list")
+    async def list_channels(self, ctx: commands.Context):
+        """
+        List blacklisted channels
+        """
+        await ctx.send(embed=discord.Embed(description="\n".join(f"``{ctx.guild.get_channel(channel)}``" for channel in self.bot.blacklist) or "None"))
 
 def setup(bot):
     bot.add_cog(ManagementCog(bot))
